requirements.txt
Flask==2.1.1
easyocr==1.5.1
pillow==8.4.0
gunicorn==20.1.0  # for production deployment
